refactor StudentProject.js
